Refactor: extract PersistentTrainer from BaseTrainer
Keep BaseTrainer clean with no persistent-specific logic.
PersistentTrainer extends BaseTrainer, overriding _train_epoch()
for chronological batching and hidden state carryover.

Addresses feedback from PR #16.
