+++ src/neuralngen/training/__init__.py
@@ -1,6 +1,8 @@
-from neuralngen.training.basetrainer import BaseTrainer, BasinChronoInterleaveSampler
+from neuralngen.training.basetrainer import BaseTrainer
+from neuralngen.training.persistent_trainer import PersistentTrainer, BasinChronoInterleaveSampler
 
 __all__ = [
     "BaseTrainer",
+    "PersistentTrainer",
     "BasinChronoInterleaveSampler",
 ]
